refactor prost hybrid into its own file
